[functorch] refactor get_fallback_and_vmap_exhaustive
This PR refactors get_fallback_and_vmap_exhaustive into two helper
functions, `generate_vmap_inputs` and
`compute_quantities_for_vmap_test`. The former is responsible for
generating in_dims and the batched inputs; the latter is responsible for
computing vmap(blah) and the fallback.

The rationale for the refactor is that I want to use these pieces
separately in in-place vmap testing, which is coming soon.

Test Plan:
- run tests

ghstack-source-id: 20a7fb23d447f40a2e0e541a9c18eed021dbc0eb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/82897